build: Add rules to build simulator executable as 'tpm2-simulator'.
Again we update the bootstrap script to generate variables holding the
paths to the source and header files for the simulator. We then update
the configure script to ensure we have pthreads available. Finally we
add rules to Makefile.am to build the simulator executable and we have
git ignore the resulting binary.

NOTE: The cyclic dependency between libplatform and libtpm requires that
we specify libplatform twice on the LDADD line for the simulator
executable.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>