ci: Use arch element to get 64bit build, make 32bit the exception.
Getting the 64bit build done through the env element is unnecessary. The
exceptional case is the 32bit build and so we use the matrix element to
get 32bit builds for clang and gcc. By having the environment clear by
default we make doing builds for other architectures possible.

This commit also reorders the elements in the compiler array to
consistently get clang builds above gcc builds in the matrix.

Finally we update the before_install element to only enable the i386
architecture and install required 32bit libs when building for 32bit.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>